requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]>=3.1
python-decouple>=3.8
django-filter>=23.5
dj-database-url>=2.1
whitenoise>=6.6
gunicorn>=21.2
Pillow>=10.2
django-extensions>=3.2
